+++ pages/genFinishedProduct.php
@@ -101,9 +101,7 @@ if($_POST['batchID'] == '1'){
                       <button type="button" class="btn btn-primary dropdown-toggle" data-bs-toggle="dropdown" aria-haspopup="true" aria-expanded="false"><i class="fa fa-bars mx-2"></i>Actions</button>
                       <div class="dropdown-menu dropdown-menu-left">
                         <li><a class="dropdown-item" id="pdf" href="#"><i class="fa-solid fa-file-pdf mx-2"></i>Export to PDF</a></li>
-                        <?php if (file_exists(__ROOT__."/pages/views/IFRA/genIFRAdoc.php")) {?>
                         <li><a class="dropdown-item" href="#" data-bs-toggle="modal" data-bs-target="#IFRA"><i class="fa-solid fa-certificate mx-2"></i>IFRA Document</a></li>
-                        <?php } ?>
                         <li><a class="dropdown-item" href="#" data-bs-toggle="modal" data-bs-target="#ViewBoxLabel"><i class="fa-solid fa-font mx-2"></i>View Box Back Label</a></li>
                       </div>
                     </div>
@@ -111,13 +109,11 @@ if($_POST['batchID'] == '1'){
                     </th>
                     </tr>
                     <tr>
-                      <th width="22%">Ingredient</th>
-                      <th width="10%">CAS#</th>
-                      <th width="10%">Purity%</th>
-                      <th width="10%">Dilutant</th>
-                      <th width="10%">Quantity</th>
-                      <th width="10%">Concentration</th>
-                      <th colspan="2">Cost</th>
+                      <th>Ingredient</th>
+                      <th>CAS#</th>
+                      <th>Quantity (<?=$settings['mUnit']?>)</th>
+                      <th>Concentration%</th>
+                      <th>Cost (<?php echo utf8_encode($settings['currency']);?>)</th>
                     </tr>
                   </thead>
                   <?php foreach ($form as $formula){
@@ -135,14 +131,9 @@ if($_POST['batchID'] == '1'){
 					?>
 					  <tr>
                       <td align="center"><?=$formula['ingredient']?></td>
-					  <td align="center"><?=$ing_q['cas']?></td>
-                      <td align="center"><?=$formula['concentration']?></td>
-					<?php if($formula['concentration'] == '100'){ ?>
-                      <td align="center">None</td>
-					<?php }else{ ?>
-					   <td data-name="dilutant" class="dilutant" data-type="select" align="center" data-pk="<?=$formula['ingredient']?>"><?=$formula['dilutant']?></td>
+					  <td align="center"><?=$ing_q['cas']?></td>					
 					 <?php
-                      }
+                      
 					  if($limit['0'] != null && $ing_q['byPassIFRA'] == 0){
 						 if($limit['0'] < $conc_p){
 							$IFRA_WARN = 'class="alert-danger"';//VALUE IS TO HIGH AGAINST IFRA
@@ -182,16 +173,12 @@ if($_POST['batchID'] == '1'){
                     </tr>
                     <tr>
                       <td></td>
-                      <td></td>
-                      <td></td>
                       <td align="center" class="m-1 text-primary">Sub Total: </td>
                       <td align="center" class="m-1 text-primary"><?php echo number_format(array_sum($new_tot), $settings['qStep']); ?> <?=$settings['mUnit']?></td>
                       <td align="center" class="m-1 text-primary"><?php echo array_sum($conc_tot);?>%</td>
                       <td colspan="2" align="center" class="m-1 text-primary"><?php echo utf8_encode($settings['currency']).number_format(array_sum($tot),$settings['qStep']);?></td>
                     </tr>
                     <tr>
-                      <td></td>
-                      <td></td>
                       <td></td>
                       <td align="center" class="m-1 text-primary">Carrier/Solvent: </td>
                       <td align="center" class="m-1 text-primary"><?php echo $carrier; ?> <?=$settings['mUnit']?></td>
@@ -200,16 +187,12 @@ if($_POST['batchID'] == '1'){
                     </tr>
                     <tr>
                       <td></td>
-                      <td></td>
-                      <td></td>
                       <td align="center" class="m-0 text-primary">Bottle:</td>
                       <td align="center" class="m-0 text-primary"><?php echo $bottle_cost['ml'];?> <?=$settings['mUnit']?></td>
                       <td align="center" class="m-0 text-primary">-</td>
                       <td colspan="2" align="center" class="m-0 text-primary"><?php echo  utf8_encode($settings['currency']).$bottle_cost['price']; ?></td>
                     </tr>
                     <tr>
-                      <td></td>
-                      <td></td>
                       <td></td>
                       <td align="center" class="m-0 text-primary">Lid:</td>
                       <td align="center" class="m-0 text-primary"><?php echo $lid_cost['style'];?></td>
@@ -218,16 +201,12 @@ if($_POST['batchID'] == '1'){
                     </tr>
                     <tr>
                       <td></td>
-                      <td></td>
-                      <td></td>
                       <td align="center" class="m-0 text-primary">Batch No:</td>
                       <td align="center" class="m-0 text-primary"><?php echo $meta['batchNo'];?></td>
                       <td align="center" class="m-0 text-primary">-</td>
                       <td colspan="2" align="center" class="m-0 text-primary">-</td>
                     </tr>
                     <tr>
-                      <td></td>
-                      <td></td>
                       <td></td>
                       <td align="center" class="m-0 font-weight-bold text-primary">Total: </td>
                       <td width="15%" align="center" class="m-0 font-weight-bold text-primary"><?php echo number_format(array_sum($new_tot)+ $carrier, $settings['qStep']); ?> <?=$settings['mUnit']?></td>
@@ -269,11 +248,8 @@ if($_POST['batchID'] == '1'){
         <h5 class="modal-title">Generate IFRA Document</h5>
       </div>
       <div class="modal-body">
-      <div class="alert alert-warning d-flex align-items-center" role="alert">
-      		IMPORTANT: The generated document isn't an official IFRA certificate and needs to be reviewed by a certified person. Also, data needs to be properly verified to make sure there are no errors.
-      </div>
+      <div class="alert alert-info"><i class="fa-solid fa-circle-exclamation mr-2"></i>IMPORTANT: The generated document isn't an official IFRA certificate and needs to be reviewed by a certified person. Also, data needs to be properly verified to make sure there are no errors.</div>
 
-      <form action="/pages/views/IFRA/genIFRAdoc.php?fid=<?php echo $meta['fid'];?>&conc=<?php echo $type; ?>&bottle=<?php echo $bottle;?>&defCatClass=<?=$defCatClass?>" method="POST" target="_blank">
 	  <div class="mb-3">
   		<label for="customer" class="form-label">Select customer</label>
         <select class="form-control" name="customer" id="customer">
@@ -300,7 +276,6 @@ if($_POST['batchID'] == '1'){
         <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Cancel</button>
         <input type="submit" class="btn btn-primary" id="generateDoc" value="Generate">
       </div>
-     </form>
     </div>
   </div>
 </div>
@@ -332,7 +307,7 @@ if($_POST['batchID'] == '1'){
    					$fTypes[] = $fTypes_res;
 				}
 			?>
-<form action="/?do=genFinishedProduct&generate=1" method="post" enctype="multipart/form-data" target="_blank">
+<form action="/?do=genFinishedProduct&generate=1" method="POST" target="_blank">
            
   <div class="row mb-3">
     <label for="formula" class="col-sm-1 col-form-label">Formula</label>
@@ -441,6 +416,27 @@ if($_POST['batchID'] == '1'){
       </div>
    </div>
   </div>
+  
+  <!-- Modal IFRA Doc -->
+  <div class="modal fade" id="ifraDocModal" data-bs-backdrop="static" tabindex="-1" role="dialog" aria-labelledby="ifraDocModalLabel" aria-hidden="true">
+  <div class="modal-dialog odal-dialog-scrollable modal-fullscreen" role="document">
+    <div class="modal-content">
+      <div class="modal-header">
+        <h5 class="modal-title" id="ifraDocModalLabel">IFRA Document</h5>
+        <button type="button" class="close" data-bs-dismiss="modal" aria-label="Close">
+          <span aria-hidden="true">&times;</span>
+        </button>
+      </div>
+      <div class="modal-body modal-IFRA-body-data">
+        Loading...
+      </div>
+      <div class="modal-footer">
+        <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
+        <button type="button" class="btn btn-primary" id="printIFRADoc">Print</button>
+      </div>
+    </div>
+  </div>
+</div>
 <script type="text/javascript" language="javascript" >
 $("#ViewBoxLabel").on("show.bs.modal", function(e) {
 	
@@ -456,6 +452,49 @@ $("#ViewBoxLabel").on("show.bs.modal", function(e) {
       $("#headerLabel", this).html("<?php echo $f_name; ?>");
       $(".modal-body", this).html(data);
     });
+	
+});
+
+$('#generateDoc').click(function() {
+	$('.modal-IFRA-body-data').html('loading');
+	$.ajax({
+	   type: 'POST',
+	   url: '/pages/views/IFRA/genIFRAdoc.php',
+	   data:{
+		   fid: "<?=$meta['fid']?>",
+		   conc: "<?=$type?>",
+		   bottle: "<?=$bottle?>",
+		   defCatClass: "<?=$defCatClass?>",
+		   template: $("#template").val(),
+		   customer: $("#customer").val()
+	   },
+	   success: function(data) {
+			$('.modal-IFRA-body-data').html(data);
+			$('#ifraDocModal').modal('show'); 
+	   },
+	   error:function(err){
+			data = '<div class="alert alert-danger">Unable to generate data</div>';
+			$('.modal-IFRA-body-data').html(data);
+			$('#ifraDocModal').modal('show'); 
+	   }
+	})
+ });
+	 
+	 
+$('#printIFRADoc').click(function() {
+    var printContents = $('.modal-IFRA-body-data').html();
+    var printWindow = window.open('', '_blank', 'width=800,height=600');
+    
+    printWindow.document.open();
+    printWindow.document.write('<html><head><title>Print</title>');
+    printWindow.document.write('</head><body>');
+    printWindow.document.write(printContents);
+    printWindow.document.write('</body></html>');
+    printWindow.document.close();
+
+    printWindow.focus();
+    printWindow.print();
+    printWindow.close();
 	
 });