+++ releasenotes.md
@@ -1,5 +1,8 @@
 Whats New in v10.7
 --------------------------
+- Removed ingredient purity and dilutant from Finished Product page for cleaner view
+- Moved IFRA doc to a modal window
+- Added a print button for the generated IFRA Doc
 - Fix IFRA Document failing to be generated
 - Added a pending materials page to list all pending materials and their quantity required for the pending formulas
 - Choose which supplier you updating the stock when making a formula